+++ include/mockturtle/algorithms/node_resynthesis/exact.hpp
@@ -41,10 +41,12 @@
 #include <kitty/dynamic_truth_table.hpp>
 #include <kitty/hash.hpp>
 #include <kitty/print.hpp>
+#include <kitty/traits.hpp>
+#include <percy/percy.hpp>
 
 #include "../../networks/aig.hpp"
+#include "../../networks/xmg.hpp"
 #include "../../networks/klut.hpp"
-#include "../../utils/include/percy.hpp"
 
 namespace mockturtle
 {
@@ -116,7 +118,7 @@ struct exact_resynthesis_params
 
    The underlying engine for this resynthesis function is percy_.
 
-   .. _percy: https://github.com/whaaswijk/percy
+   .. _percy: https://github.com/lsils/percy
    \endverbatim
  *
  */
@@ -231,7 +233,7 @@ private:
  *
  * This resynthesis function can be passed to ``node_resynthesis``,
  * ``cut_rewriting``, and ``refactoring``.  The given truth table will be
- * resynthized in terms of an optimum size AIG network.
+ * resynthesized in terms of an optimum size AIG network.
  *
    \verbatim embed:rst
 
@@ -239,7 +241,7 @@ private:
 
    .. code-block:: c++
 
-      const aig_network aig = ...;
+      aig_network aig = ...;
 
       exact_aig_resynthesis<aig_network> resyn;
       aig = cut_rewriting( aig, resyn );
@@ -256,7 +258,7 @@ private:
 
    .. code-block:: c++
 
-      const aig_network aig = ...;
+      aig_network aig = ...;
 
       exact_resynthesis_params ps;
       ps.cache = std::make_shared<exact_resynthesis_params::cache_map_t>();
@@ -265,7 +267,7 @@ private:
 
    The underlying engine for this resynthesis function is percy_.
 
-   .. _percy: https://github.com/whaaswijk/percy
+   .. _percy: https://github.com/lsils/percy
    \endverbatim
  *
  */
@@ -348,7 +350,6 @@ public:
     }
 
     std::vector<signal<Ntk>> signals( begin, end );
-
     for ( auto i = 0; i < c->get_nr_steps(); ++i )
     {
       auto c1 = signals[c->get_step( i )[0]];
@@ -393,5 +394,139 @@ private:
   std::optional<uint32_t> _lower_bound;
   std::optional<uint32_t> _upper_bound;
 };
+
+struct exact_xmg_resynthesis_params
+{
+  uint32_t num_candidates{10u};
+};
+
+/*! \brief Resynthesis function based on exact synthesis for XMGs.
+ *
+ * This resynthesis function can be passed to ``node_resynthesis``,
+ * ``cut_rewriting``, and ``refactoring``.  The given truth table will be
+ * resynthesized in terms of an optimum size XMG network.
+ *
+   \verbatim embed:rst
+
+   Example
+
+   .. code-block:: c++
+
+      xmg_network aig = ...;
+
+      exact_xmg_resynthesis<xmg_network> resyn;
+      xmg = cut_rewriting( xmg, resyn );
+   \endverbatim
+ *
+ *
+   The underlying engine for this resynthesis function is percy_.
+
+   .. _percy: https://github.com/lsils/percy
+   \endverbatim
+ *
+ */
+template<class Ntk = xmg_network>
+class exact_xmg_resynthesis
+{
+public:
+  explicit exact_xmg_resynthesis( exact_xmg_resynthesis_params const& ps = {} )
+    : ps( ps )
+  {
+  }
+
+  template<typename LeavesIterator, typename TT, typename Fn>
+  void operator()( Ntk& ntk, TT const& function, LeavesIterator begin, LeavesIterator end, Fn&& fn ) const
+  {
+    static_assert( kitty::is_complete_truth_table<TT>::value, "Truth table must be complete" );
+
+    using signal = mockturtle::signal<Ntk>;
+
+    percy::chain chain;
+    percy::spec spec;
+    spec.verbosity = 0;
+    spec.fanin = 3;
+
+    /* specify local normalized gate primitives */
+    kitty::dynamic_truth_table const0{3};
+    kitty::dynamic_truth_table a{3};
+    kitty::dynamic_truth_table b{3};
+    kitty::dynamic_truth_table c{3};
+    kitty::create_nth_var( a, 0 );
+    kitty::create_nth_var( b, 1 );
+    kitty::create_nth_var( c, 2 );
+
+    spec.add_primitive( const0 );
+    spec.add_primitive( a );
+    spec.add_primitive( b );
+    spec.add_primitive( c );
+    spec.add_primitive( kitty::ternary_majority(  a,  b,  c ) );
+    spec.add_primitive( kitty::ternary_majority( ~a,  b,  c ) );
+    spec.add_primitive( kitty::ternary_majority(  a, ~b,  c ) );
+    spec.add_primitive( kitty::ternary_majority(  a,  b, ~c ) );
+    spec.add_primitive( a ^ b ^ c );
+
+    percy::bsat_wrapper solver;
+    percy::ssv_encoder encoder(solver);
+
+    spec[0] = kitty::is_normal( function ) ? function : ~function;
+
+    for ( auto i = 0u; i < ps.num_candidates; ++i )
+    {
+      auto const result = percy::next_struct_solution( spec, chain, solver, encoder );
+      if ( result != percy::success )
+        break;
+
+      assert( result == percy::success );
+
+      auto const sim = chain.simulate();
+      assert( chain.simulate()[0] == spec[0] );
+
+      std::vector<signal> signals( begin, end );
+      for ( auto i = 0; i < chain.get_nr_steps(); ++i )
+      {
+        auto const c1 = signals[chain.get_step( i )[0]];
+        auto const c2 = signals[chain.get_step( i )[1]];
+        auto const c3 = signals[chain.get_step( i )[2]];
+
+        switch( chain.get_operator( i )._bits[0] )
+        {
+        case 0x00:
+          signals.emplace_back( ntk.get_constant( false ) );
+          break;
+
+        case 0xe8:
+          signals.emplace_back( ntk.create_maj( c1,  c2,  c3 ) );
+          break;
+
+        case 0xd4:
+          signals.emplace_back( ntk.create_maj( !c1,  c2,  c3 ) );
+          break;
+
+        case 0xb2:
+          signals.emplace_back( ntk.create_maj( c1,  !c2,  c3 ) );
+          break;
+
+        case 0x8e:
+          signals.emplace_back( ntk.create_maj( c1,  c2,  !c3 ) );
+          break;
+
+        case 0x96:
+          signals.emplace_back( ntk.create_xor3( c1,  c2,  c3 ) );
+          break;
+
+        default:
+          std::cerr << "[e] unsupported operation " << kitty::to_hex( chain.get_operator( i ) ) << "\n";
+          assert( false );
+          break;
+        }
+      }
+
+      fn( chain.is_output_inverted( 0 ) ? !signals.back() : signals.back() );
+    }
+  }
+
+protected:
+  exact_xmg_resynthesis_params const& ps;
+}; /* exact_xmg_resynthesis */
 
 } /* namespace mockturtle */